Replace eclipse logo with our the new Eclipse logo.
An hi-res version of the new logo is smaller than a low-res version of
our old logo. We can now support retina display and still save on
bandwidth/pageload time.
    
This patch replace eclipse.png default logo with the new logo using the
same dimensions. This will avoid breaking the layout of sites that are
using a copy of our logo with a direct link to our servers. This version
does not support retina display.
    
I also created 2 other versions that support retina display. One include
white space so that it can be inserted to Nova without any changes to
the CSS.
    
Signed-off-by: Christopher Guindon <chris.guindon@eclipse.org>